(Update): Removed all commands syntax outside to another class
